Removing i_cpu and affinity stuff and adding ability to reset GL_Init to apply some settings while in-game

+++ src/engine/i_swap.h
@@ -32,7 +32,7 @@
 #define I_SwapBE32(x)   SDL_SwapBE32(x)
 
 #define SHORT(x)        ((signed short)I_SwapLE16(x))
-#define LONG(x)         ((signed long)I_SwapLE32(x))
+#define LONG(x)         ((signed long long)I_SwapLE32(x))
 
 // Defines for checking the endianness of the system.
